Struct ogg_magic::ogg_vorbis_page::OggVorbisPage

pub struct OggVorbisPage {
    pub ogg_page: OggPage,
}

Represents an Ogg Vorbis page.

Fields
ogg_page: OggPage
Implementations

impl OggVorbisPage

pub fn new(buffer: Vec<u8>) -> Result<Self, VorbisFormatError>

Creates a new OggVorbisPage from the given buffer.

pub fn get_identification( &self, segment_index: usize, ) -> Result<IVorbisIdentificationHeader, VorbisFormatError>

Retrieves the identification header from the specified segment index.

pub fn is_header_packet(&self, segment_index: usize) -> bool

Checks if the specified segment is a header packet.

pub fn is_identification_packet(&self, segment_index: usize) -> bool

Checks if the specified segment is an identification packet.

pub fn is_comment_packet(&self, segment_index: usize) -> bool

Checks if the specified segment is a comment packet.

pub fn is_setup_packet(&self, segment_index: usize) -> bool

Checks if the specified segment is a setup packet.

pub fn get_comments( &self, segment_index: usize, ) -> Result<IVorbisCommentHeader, VorbisFormatError>

Retrieves the comments from the specified segment index.

pub fn get_setup( &self, audio_channels: u8, segment_index: usize, ) -> Result<IVorbisSetupHeader, VorbisFormatError>

Retrieves the setup header from the specified segment index.

pub fn remove_page_segment( &self, index: usize, n: usize, ) -> Result<Self, VorbisFormatError>

pub fn replace_page_segment( &self, segment: Vec<u8>, index: usize, ) -> Result<Self, VorbisFormatError>

pub fn build_comments(header: IVorbisCommentHeader) -> Vec<u8>