Module ogg_magic::read_ogg_vorbis_file

Structs

OggParseResult

Represents the result of parsing an Ogg page, containing the type of result, the parsed data, the index of the segment, and the raw segment data.

OggVorbisPageResult

Represents the result of parsing an Ogg Vorbis page, containing the page and its associated packets.

Enums

OggVorbisPacket

Enum representing different types of Ogg Vorbis packets.

Functions

read_ogg_vorbis_file

Reads an Ogg Vorbis file asynchronously and parses its pages and packets.

Type Aliases

OggVorbisPacketsParseResult

Type alias for the result of parsing Ogg Vorbis packets.

OggVorbiseHeaderCommentParseResult

Type alias for the result of parsing an Ogg Vorbis comment header.

OggVorbiseHeaderIdentificationParseResult

Type alias for the result of parsing an Ogg Vorbis identification header.

OggVorbiseHeaderSetupParseResult

Type alias for the result of parsing an Ogg Vorbis setup header.